+++ src/index.ts
@@ -3,6 +3,7 @@ export { default as distributeValue } from './distributeValue';
 export { default as googleAds } from './InteractionMiddlewares/GoogleAds';
 export { default as facebookAds } from './InteractionMiddlewares/FacebookAds';
 export { default as ref } from './InteractionMiddlewares/Ref';
+export { default as ReferralMapper } from './InteractionMiddlewares/ReferralMapper';
 export { default as lastInteraction } from './AttributionModels/LastInteraction';
 export { default as firstInteraction } from './AttributionModels/FirstInteraction';
 export { default as lastNonDirectInteraction } from './AttributionModels/LastNonDirectInteraction';